chore: Move runtime components out of 'js-runtime' and rename package to 'js-compiler'
