+++ packages/generated-patterns/integration/pattern-harness.ts
@@ -1,7 +1,7 @@
 import { expect } from "@std/expect";
 import "@commontools/utils/equal-ignoring-symbols";
 import { fromFileUrl } from "@std/path";
-import { FileSystemProgramResolver } from "@commontools/js-runtime";
+import { FileSystemProgramResolver } from "@commontools/js-compiler";
 import { Identity } from "@commontools/identity";
 import { StorageManager } from "../../runner/src/storage/cache.deno.ts";
 import { Runtime } from "@commontools/runner";
