+++ packages/js-compiler/typescript/bundler/amd-loader.ts
