--- packages/js-runtime/runtime/mod.ts
@@ -1,1 +1,0 @@
-export * from "./eval-runtime.ts";
